lib/ukintctlr: Migrate PIC implementation to the xPIC driver
Migrate PIC to drivers/ukintctlr/xpic. Move architectural definitions
to arch/.

Notice: Picking individual commits in this PR will break the build.

Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1103
