+++ arch/x86/x86_64/include/uk/asm/arch.h
@@ -1,0 +1,11 @@
+/* SPDX-License-Identifier: BSD-3-Clause */
+/* Copyright (c) 2023, Unikraft GmbH and The Unikraft Authors.
+ * Licensed under the BSD-3-Clause License (the "License").
+ * You may not use this file except in compliance with the License.
+ */
+#ifndef __UK_ARCH_X86_64_H__
+#define __UK_ARCHX86_64__H__
+
+#include <uk/asm/pic.h>
+
+#endif /* __UK_ARCH_X86_64_H__ */
