+++ arch/x86/x86_64/include/uk/asm/pic.h
@@ -22,9 +22,8 @@
  */
 /* Taken from solo5 platform_intr.c */
 
-#include <stdint.h>
-#include <x86/cpu.h>
-#include <uk/intctlr.h>
+#ifndef __UK_ARCH_X86_64_PIC_H__
+#define __UK_ARCH_X86_64_PIC_H__
 
 #define PIC1             0x20    /* IO base address for master PIC */
 #define PIC2             0xA0    /* IO base address for slave PIC */
@@ -48,64 +47,4 @@
 #define ICW4_BUF_MASTER  0x0C /* Buffered mode/master */
 #define ICW4_SFN         0x10 /* Special fully nested (not) */
 
-/*
- * arguments:
- * offset1 - vector offset for master PIC vectors on the master become
- *           offset1..offset1+7
- * offset2 - same for slave PIC: offset2..offset2+7
- */
-static void PIC_remap(int offset1, int offset2)
-{
-	unsigned char a1, a2;
-
-	/* save masks */
-	a1 = inb(PIC1_DATA);
-	a2 = inb(PIC2_DATA);
-
-	/* start init seq (cascade) */
-	outb(PIC1_COMMAND, ICW1_INIT + ICW1_ICW4);
-	outb(PIC2_COMMAND, ICW1_INIT + ICW1_ICW4);
-	/* ICW2: Master PIC vector off */
-	outb(PIC1_DATA, offset1);
-	/* ICW2: Slave PIC vector off */
-	outb(PIC2_DATA, offset2);
-	/* ICW3: tell Master PIC there is a slave PIC at IRQ2 (0000 0100) */
-	outb(PIC1_DATA, 4);
-	/* ICW3: tell Slave PIC its cascade identity (0000 0010) */
-	outb(PIC2_DATA, 2);
-
-	outb(PIC1_DATA, ICW4_8086);
-	outb(PIC2_DATA, ICW4_8086);
-
-	outb(PIC1_DATA, a1); /* restore saved masks. */
-	outb(PIC2_DATA, a2);
-}
-
-void intctrl_init(void)
-{
-	PIC_remap(32, 40);
-}
-
-void intctrl_ack_irq(unsigned int irq)
-{
-	if (!IRQ_ON_MASTER(irq))
-		outb(PIC2_COMMAND, PIC_EOI);
-
-	outb(PIC1_COMMAND, PIC_EOI);
-}
-
-void intctrl_mask_irq(unsigned int irq)
-{
-	__u16 port;
-
-	port = IRQ_PORT(irq);
-	outb(port, inb(port) | (1 << IRQ_OFFSET(irq)));
-}
-
-void intctrl_clear_irq(unsigned int irq)
-{
-	__u16 port;
-
-	port = IRQ_PORT(irq);
-	outb(port, inb(port) & ~(1 << IRQ_OFFSET(irq)));
-}
+#endif /* __UK_ARCH_X86_64_PIC_H__ */
