+++ drivers/ukintctlr/xpic/Makefile.uk
@@ -3,3 +3,7 @@ $(eval $(call addlib_s,libukintctlr_xpic,$(CONFIG_LIBUKINTCTLR_XPIC)))
 ASINCLUDES-$(CONFIG_LIBUKINTCTLR_XPIC)  += -I$(LIBUKINTCTLR_XPIC_BASE)/include
 CINCLUDES-$(CONFIG_LIBUKINTCTLR_XPIC)   += -I$(LIBUKINTCTLR_XPIC_BASE)/include
 CXXINCLUDES-$(CONFIG_LIBUKINTCTLR_XPIC) += -I$(LIBUKINTCTLR_XPIC_BASE)/include
+
+LIBUKINTCTLR_XPIC_CINCLUDES-y += -I$(CONFIG_UK_BASE)/plat/common/include
+
+LIBUKINTCTLR_XPIC_SRCS-y += $(LIBUKINTCTLR_XPIC_BASE)/pic.c
