+++ lib/ukintctlr/Makefile.uk
@@ -4,12 +4,4 @@ ASINCLUDES-$(CONFIG_LIBUKINTCTLR)  += -I$(LIBUKINTCTLR_BASE)/include
 CINCLUDES-$(CONFIG_LIBUKINTCTLR)   += -I$(LIBUKINTCTLR_BASE)/include
 CXXINCLUDES-$(CONFIG_LIBUKINTCTLR) += -I$(LIBUKINTCTLR_BASE)/include
 
-ASINCLUDES-$(CONFIG_LIBUKINTCTLR)  += -I$(CONFIG_UK_BASE)/plat/common/include
-CINCLUDES-$(CONFIG_LIBUKINTCTLR)   += -I$(CONFIG_UK_BASE)/plat/common/include
-CXXINCLUDES-$(CONFIG_LIBUKINTCTLR) += -I$(CONFIG_UK_BASE)/plat/common/include
-
-ifeq ($(CONFIG_ARCH_X86_64),y)
-LIBUKINTCTLR_SRCS-y += $(LIBUKINTCTLR_BASE)/arch/x86/intctlr.c
-endif
-
 LIBUKINTCTLR_SRCS-y += $(LIBUKINTCTLR_BASE)/ukintctlr.c
