+++ lib/ukintctlr/exportsyms.uk
@@ -1,8 +1,3 @@
-intctrl_init
-intctrl_clear_irq
-intctrl_mask_irq
-intctrl_ack_irq
-intctrl_send_ipi
 uk_intctlr
 uk_intctlr_init
 uk_intctlr_irq_configure
